requirements.txt
# 核心依赖
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.2

# 机器学习
xgboost>=2.0.0
shap>=0.42.0
optuna>=3.0.0

# Web框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# 数据库
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pymongo>=4.5.0

# 工具
python-dotenv>=1.0.0
